requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
azure-cognitiveservices-speech>=1.31.0
azure-ai-textanalytics>=5.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic==2.4.2
openai-whisper==20231117
pyaudio==0.2.13
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0 